# Makes throttling parameters obvious by controlling exceptions