--- a/zipkin-server/src/main/java/zipkin2/server/internal/throttle/ThrottledStorageComponent.java
+++ b/zipkin-server/src/main/java/zipkin2/server/internal/throttle/ThrottledStorageComponent.java
@@ -22,6 +22,7 @@ import com.netflix.concurrency.limits.Limiter;
 import com.netflix.concurrency.limits.limit.Gradient2Limit;
 import com.netflix.concurrency.limits.limiter.AbstractLimiter;
 import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.util.NamedThreadFactory;
 import java.io.IOException;
 import java.util.List;
 import java.util.Objects;
@@ -29,7 +30,7 @@ import java.util.Optional;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.Executor;
 import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
@@ -41,6 +42,8 @@ import zipkin2.server.internal.brave.TracedCall;
 import zipkin2.storage.ForwardingStorageComponent;
 import zipkin2.storage.SpanConsumer;
 import zipkin2.storage.StorageComponent;
+
+import static com.linecorp.armeria.common.util.Exceptions.clearTrace;
 
 /**
  * Delegating implementation that limits requests to the {@link #spanConsumer()} of another {@link
@@ -56,6 +59,13 @@ import zipkin2.storage.StorageComponent;
  * @see ThrottledSpanConsumer
  */
 public final class ThrottledStorageComponent extends ForwardingStorageComponent {
+  /**
+   * Rather than flooding when queue size reached, return the same instance. The path to this is
+   * unimportant, so we clear the trace.
+   */
+  static final RejectedExecutionException STORAGE_THROTTLE_MAX_QUEUE_SIZE =
+    clearTrace(new RejectedExecutionException("STORAGE_THROTTLE_MAX_QUEUE_SIZE reached"));
+
   final StorageComponent delegate;
   final @Nullable Tracing tracing;
   final AbstractLimiter<Void> limiter;
@@ -76,15 +86,18 @@ public final class ThrottledStorageComponent extends ForwardingStorageComponent 
       .build();
     this.limiter = new Builder().limit(limit).build();
 
-    // TODO: explain these parameters
-    executor = new ThreadPoolExecutor(limit.getLimit(),
+    // The size of the thread pool is managed by the limiter, so we initialize it with the lower
+    // bound (current limit), and later use change notification to resize it.
+    executor = new ThreadPoolExecutor(
+      limit.getLimit(),
       limit.getLimit(),
       0,
       TimeUnit.DAYS,
-      createQueue(maxQueueSize),
-      new ThrottledThreadFactory(),
-      new ThreadPoolExecutor.AbortPolicy());
-
+      createQueue(maxQueueSize), //
+      new NamedThreadFactory("zipkin-throttle-pool"),
+      (r, e) -> {
+        throw STORAGE_THROTTLE_MAX_QUEUE_SIZE;
+      });
     limit.notifyOnChange(new ThreadPoolExecutorResizer(executor));
 
     ActuateThrottleMetrics metrics = new ActuateThrottleMetrics(registry);
@@ -176,18 +189,6 @@ public final class ThrottledStorageComponent extends ForwardingStorageComponent 
     }
 
     return new LinkedBlockingQueue<>(maxSize);
-  }
-
-  static final class ThrottledThreadFactory implements ThreadFactory {
-    ThrottledThreadFactory() {
-    }
-
-    @Override public Thread newThread(Runnable r) {
-      Thread thread = new Thread(r);
-      thread.setDaemon(true);
-      thread.setName("zipkin-throttle-pool-" + thread.getId());
-      return thread;
-    }
   }
 
   static final class ThreadPoolExecutorResizer implements Consumer<Integer> {