@adriancole 2019-08-08
I guess my intent was that there is only one place the exception originates
from, and its presence isnt impacted based on what the call path to it is..
hence the stack being uninteresting and certainly not interesting enough to
make lots of logs when you try to look what's happening.

otoh maybe it isnt important enough to devise such an explanation ;)

On Thu, Aug 8, 2019, 2:09 PM Anuraag Agrawal <notifications@github.com>
wrote:

> *@anuraaga* approved this pull request.
> ------------------------------
>
> In
> zipkin-server/src/main/java/zipkin2/server/internal/throttle/ThrottledCall.java
> <https://github.com/openzipkin/zipkin/pull/2742#discussion_r311867891>:
>
> > @@ -38,6 +40,13 @@
>   * @see ThrottledStorageComponent
>   */
>  final class ThrottledCall extends Call.Base<Void> {
> +  /**
> +   * Rather than flooding when concurrency reached, return the same instance. The path to this is
>
> I'm not sure what flooding means here, so the comment takes much longer to
> parse. Is it more like Reduce allocations when concurrency reached by
> always returning the same instance. This is only thrown in one location so
> the stack trace isn't useful and we clear it to keep logs down.?
> ------------------------------
>
> In
> zipkin-server/src/main/java/zipkin2/server/internal/throttle/ThrottledCall.java
> <https://github.com/openzipkin/zipkin/pull/2742#discussion_r311868538>:
>
> > @@ -38,6 +40,13 @@
>   * @see ThrottledStorageComponent
>   */
>  final class ThrottledCall extends Call.Base<Void> {
> +  /**
> +   * Rather than flooding when concurrency reached, return the same instance. The path to this is
> +   * unimportant, so we clear the trace.
> +   */
> +  static final RejectedExecutionException STORAGE_THROTTLE_MAX_CONCURRENCY =
> +    clearTrace(new RejectedExecutionException("STORAGE_THROTTLE_MAX_CONCURRENCY reached"));
>
> Maybe we can add a link to a javadoc or something to these trace-cleared
> exceptions to give users something to check if they've never seen the
> exception before.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/openzipkin/zipkin/pull/2742?email_source=notifications&email_token=AAAPVV5UG2EH5NYYY2CH733QDO2C7A5CNFSM4IKGVFUKYY3PNVWWK3TUL52HS4DFWFIHK3DMKJSXC5LFON2FEZLWNFSXPKTDN5WW2ZLOORPWSZGOCA55WNQ#pullrequestreview-272358198>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAAPVV3SB7NKXVUO3LKUG4LQDO2C7ANCNFSM4IKGVFUA>
> .
>