@adriancole 2019-08-08
here's the impact. Notably in the throttled case, there's an order of magnitude less allocation due to the exception sharing.

before
ThrottledCallBenchmarks.execute_throttled:·gc.alloc.rate.norm               sample      15    792.430 ±   0.169    B/op

after
ThrottledCallBenchmarks.execute_throttled:·gc.alloc.rate.norm              sample      15    72.005 ±   0.001    B/op

whole thing below:

before:

```
Benchmark                                                                     Mode     Cnt      Score     Error   Units
ThrottledCallBenchmarks.execute                                             sample  553318     15.337 ±   0.050   us/op
ThrottledCallBenchmarks.execute:execute·p0.00                               sample              8.352             us/op
ThrottledCallBenchmarks.execute:execute·p0.50                               sample             13.696             us/op
ThrottledCallBenchmarks.execute:execute·p0.90                               sample             20.224             us/op
ThrottledCallBenchmarks.execute:execute·p0.95                               sample             23.808             us/op
ThrottledCallBenchmarks.execute:execute·p0.99                               sample             35.584             us/op
ThrottledCallBenchmarks.execute:execute·p0.999                              sample             66.816             us/op
ThrottledCallBenchmarks.execute:execute·p0.9999                             sample            117.632             us/op
ThrottledCallBenchmarks.execute:execute·p1.00                               sample           3846.144             us/op
ThrottledCallBenchmarks.execute:·gc.alloc.rate                              sample      15     34.036 ±   6.975  MB/sec
ThrottledCallBenchmarks.execute:·gc.alloc.rate.norm                         sample      15    399.731 ±  60.318    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Eden_Space                     sample      15     32.786 ±  51.311  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Eden_Space.norm                sample      15    376.350 ± 593.194    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Old_Gen                        sample      15      0.042 ±   0.093  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Old_Gen.norm                   sample      15      0.506 ±   1.162    B/op
ThrottledCallBenchmarks.execute:·gc.count                                   sample      15      5.000            counts
ThrottledCallBenchmarks.execute:·gc.time                                    sample      15     12.000                ms
ThrottledCallBenchmarks.execute_overCapacity                                sample  569158     15.863 ±   0.166   us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.00     sample              6.856             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.50     sample             13.776             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.90     sample             20.544             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.95     sample             27.008             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.99     sample             45.184             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.999    sample             71.424             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.9999   sample            203.415             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p1.00     sample          19333.120             us/op
ThrottledCallBenchmarks.execute_overCapacity:·gc.alloc.rate                 sample      15     32.967 ±   5.845  MB/sec
ThrottledCallBenchmarks.execute_overCapacity:·gc.alloc.rate.norm            sample      15    399.932 ±  60.440    B/op
ThrottledCallBenchmarks.execute_overCapacity:·gc.churn.G1_Eden_Space        sample      15     39.279 ±  53.240  MB/sec
ThrottledCallBenchmarks.execute_overCapacity:·gc.churn.G1_Eden_Space.norm   sample      15    490.696 ± 670.245    B/op
ThrottledCallBenchmarks.execute_overCapacity:·gc.churn.G1_Old_Gen           sample      15      0.054 ±   0.156  MB/sec
ThrottledCallBenchmarks.execute_overCapacity:·gc.churn.G1_Old_Gen.norm      sample      15      0.656 ±   1.915    B/op
ThrottledCallBenchmarks.execute_overCapacity:·gc.count                      sample      15      6.000            counts
ThrottledCallBenchmarks.execute_overCapacity:·gc.time                       sample      15     32.000                ms
ThrottledCallBenchmarks.execute_throttled                                   sample  697600      1.982 ±   0.481   us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.00           sample              1.052             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.50           sample              1.464             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.90           sample              1.744             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.95           sample              2.964             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.99           sample              5.552             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.999          sample             22.477             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.9999         sample             89.815             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p1.00           sample          84803.584             us/op
ThrottledCallBenchmarks.execute_throttled:·gc.alloc.rate                    sample      15    638.903 ±  78.012  MB/sec
ThrottledCallBenchmarks.execute_throttled:·gc.alloc.rate.norm               sample      15    792.430 ±   0.169    B/op
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Eden_Space           sample      15    644.752 ±  93.017  MB/sec
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Eden_Space.norm      sample      15    801.844 ±  87.342    B/op
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Old_Gen              sample      15      0.096 ±   0.196  MB/sec
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Old_Gen.norm         sample      15      0.126 ±   0.273    B/op
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Survivor_Space       sample      15      0.177 ±   0.324  MB/sec
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Survivor_Space.norm  sample      15      0.221 ±   0.411    B/op
ThrottledCallBenchmarks.execute_throttled:·gc.count                         sample      15     76.000            counts
ThrottledCallBenchmarks.execute_throttled:·gc.time                          sample      15     64.000                ms
```

after:

```
Benchmark                                                                    Mode     Cnt     Score     Error   Units
ThrottledCallBenchmarks.execute                                            sample  584455    16.308 ±   0.085   us/op
ThrottledCallBenchmarks.execute:execute·p0.00                              sample             8.016             us/op
ThrottledCallBenchmarks.execute:execute·p0.50                              sample            14.000             us/op
ThrottledCallBenchmarks.execute:execute·p0.90                              sample            22.592             us/op
ThrottledCallBenchmarks.execute:execute·p0.95                              sample            26.912             us/op
ThrottledCallBenchmarks.execute:execute·p0.99                              sample            39.040             us/op
ThrottledCallBenchmarks.execute:execute·p0.999                             sample            78.475             us/op
ThrottledCallBenchmarks.execute:execute·p0.9999                            sample           482.182             us/op
ThrottledCallBenchmarks.execute:execute·p1.00                              sample          3874.816             us/op
ThrottledCallBenchmarks.execute:·gc.alloc.rate                             sample      15    32.418 ±   6.469  MB/sec
ThrottledCallBenchmarks.execute:·gc.alloc.rate.norm                        sample      15   400.573 ±  60.184    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Eden_Space                    sample      15    32.288 ±  50.541  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Eden_Space.norm               sample      15   412.093 ± 650.973    B/op
ThrottledCallBenchmarks.execute:·gc.churn.G1_Old_Gen                       sample      15     0.046 ±   0.107  MB/sec
ThrottledCallBenchmarks.execute:·gc.churn.G1_Old_Gen.norm                  sample      15     0.584 ±   1.324    B/op
ThrottledCallBenchmarks.execute:·gc.count                                  sample      15     5.000            counts
ThrottledCallBenchmarks.execute:·gc.time                                   sample      15    44.000                ms
ThrottledCallBenchmarks.execute_overCapacity                               sample  479241    15.658 ±   0.051   us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.00    sample             7.216             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.50    sample            14.000             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.90    sample            19.872             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.95    sample            24.096             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.99    sample            41.152             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.999   sample            78.049             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p0.9999  sample           136.487             us/op
ThrottledCallBenchmarks.execute_overCapacity:execute_overCapacity·p1.00    sample          2756.608             us/op
ThrottledCallBenchmarks.execute_overCapacity:·gc.alloc.rate                sample      15    32.762 ±   5.522  MB/sec
ThrottledCallBenchmarks.execute_overCapacity:·gc.alloc.rate.norm           sample      15   400.054 ±  60.163    B/op
ThrottledCallBenchmarks.execute_overCapacity:·gc.churn.G1_Eden_Space       sample      15    39.179 ±  53.105  MB/sec
ThrottledCallBenchmarks.execute_overCapacity:·gc.churn.G1_Eden_Space.norm  sample      15   482.943 ± 655.319    B/op
ThrottledCallBenchmarks.execute_overCapacity:·gc.churn.G1_Old_Gen          sample      15     0.061 ±   0.146  MB/sec
ThrottledCallBenchmarks.execute_overCapacity:·gc.churn.G1_Old_Gen.norm     sample      15     0.760 ±   1.820    B/op
ThrottledCallBenchmarks.execute_overCapacity:·gc.count                     sample      15     6.000            counts
ThrottledCallBenchmarks.execute_overCapacity:·gc.time                      sample      15    17.000                ms
ThrottledCallBenchmarks.execute_throttled                                  sample  853877     0.114 ±   0.004   us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.00          sample             0.001             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.50          sample             0.088             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.90          sample             0.117             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.95          sample             0.160             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.99          sample             0.438             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.999         sample             3.505             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p0.9999        sample            13.930             us/op
ThrottledCallBenchmarks.execute_throttled:execute_throttled·p1.00          sample          1001.472             us/op
ThrottledCallBenchmarks.execute_throttled:·gc.alloc.rate                   sample      15  3724.791 ± 185.890  MB/sec
ThrottledCallBenchmarks.execute_throttled:·gc.alloc.rate.norm              sample      15    72.005 ±   0.001    B/op
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Eden_Space          sample      15  3738.655 ± 230.096  MB/sec
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Eden_Space.norm     sample      15    72.269 ±   2.414    B/op
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Old_Gen             sample      15     0.011 ±   0.005  MB/sec
ThrottledCallBenchmarks.execute_throttled:·gc.churn.G1_Old_Gen.norm        sample      15    ≈ 10⁻⁴              B/op
ThrottledCallBenchmarks.execute_throttled:·gc.count                        sample      15   208.000            counts
ThrottledCallBenchmarks.execute_throttled:·gc.time                         sample      15   162.000                ms
```